GridTheory
Tired of the aesthetic use/abuse of Bootstrap? Foundation lacking documentation for that specific question you want to drill down to?

Other Responsive and Fluid Frameworks exist, but who has the time to search + test + experiment? Oh, and you get extra files and docs to mull through. Congrats!

GridTheory isn't a Framework. It's a Grid system. Pure. Simple. Easy. That, and it's entirely structured via CSS. Intended for mobile-first design. If that's not your gig, it's an easy fix.

Tweaking to your needs is as simple as 1,2,3.
1.) Find your intended element/selector (default are Containers and Divs, which create Columns and Rows)
2.) Edit your intended element/selector. Optionally, edit the default settings per each element/selector.
3.) Oh..wait...you're done.

Looks like only two steps. That said, this is a work in progress. I am confident enough to post an "Alpha" now.

Intended to be the "Normalize.css" of Grid stylesheets.

That being said, here's a pro tip:
The big boy frameworks like Bootstrap and Foundation are built using a simple CSS Grid system. The "Framework" title comes from the extras you get on top. Icons, Fonts, Icon Fonts, Color Scheme, jQuery Snippets, and tons of other little goodies.

This has none of that. Maybe later, and it will be a separate project, but for now this is intended as a damn simple Grid System. 

-Cheers!